refactor(kolme): extract api codec module ownership (#1956)
