+++ crates/kamn-core/tests/kolme_runtime_commit_extraction_plan_docs.rs
@@ -77,6 +77,7 @@ fn regression_requires_phase_gates_and_validation_matrix_markers() {
     assert!(DOC.contains("#1950"));
     assert!(DOC.contains("#1952"));
     assert!(DOC.contains("#1954"));
+    assert!(DOC.contains("#1956"));
     assert!(DOC.contains("## Validation Matrix"));
     assert!(DOC.contains("Regression: #1814"));
 }
