+++ docs/planning/kolme_runtime_commit_extraction_plan.md
@@ -95,6 +95,7 @@ Out of scope:
 - #1950: extracted adapter-backed client ownership into `kolme_runtime_commit/adapter_backed_client.rs` and rewired `kolme_runtime_commit.rs` to re-export `AdapterBackedKolmeRuntimeCommitClient` from the dedicated submodule.
 - #1952: extracted live provider ownership into `kolme_runtime_commit/live_provider.rs` and rewired `kolme_runtime_commit.rs` to re-export `KolmeRuntimeCommitLiveProvider` from the dedicated submodule.
 - #1954: extracted HTTP transport ownership into `kolme_runtime_commit/http_transport.rs` and rewired `kolme_runtime_commit.rs` to re-export `KolmeRuntimeCommitHttpTransport` from the dedicated submodule.
+- #1956: extracted API codec ownership into `kolme_runtime_commit/api_codec.rs` and rewired `kolme_runtime_commit.rs` to re-export `KolmeApiNextNonce*`/`KolmeApiBroadcast*` types from the dedicated submodule.
 
 ## Phase 3 - Adapter and lifecycle orchestration extraction
 - split remaining adapter/transport bridge glue into dedicated modules (or subcrate) with explicit ownership,
